perf(decode): fuse match-copy into VBMI2 + BMI2 seq monoliths
The AVX2-tier exec fuse (call-boundary removal via macro) only covered the
AVX2 kernel; VBMI2 and BMI2 tiers still called the exec trait method per
sequence. Share the exec body as two macros in exec_sequence_inline:

- exec_sequence_avx2_inline (32-byte ymm, offset>=32 path) — AVX2 + VBMI2
  (VBMI2 always implies AVX2+BMI2)
- exec_sequence_sse2_inline (16-byte xmm, overshoot 15) — BMI2 (no AVX2)

Both expand textually at the sequence-loop call site so the match-copy
fuses into each per-tier monolith (target_feature fns can't be
inline(always), rust#145574). The exec_sequence_inline / _avx2 trait
methods remain as the unit-tested reference spec the macros mirror.

VBMI2/BMI2 paths are not reachable on the available test hardware
(i9-9900K = AVX2 tier, no AVX-512; M1 = aarch64; i686 = scalar): the
VBMI2 macro is identical to the i9-verified AVX2 expansion, and the BMI2
SSE2 macro mirrors the directly unit-tested exec_sequence_inline body.
